+++ api/.env
@@ -1,0 +1,2 @@
+GCP_PROJECT_ID = "hackathon-2025-455513"
+GCP_TOPIC_ID = "guardian"
