+++ api/requirements.txt
@@ -1,0 +1,5 @@
+Flask
+ipaddress
+google-cloud-pubsub
+python-dotenv
+
